+++ feature-mypage/src/main/java/com/materip/feature_mypage/screen/MyPage/ProfileMainContent.kt
@@ -1,4 +1,4 @@
-package com.materip.feature_mypage.screen
+package com.materip.feature_mypage.screen.MyPage
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
@@ -97,7 +97,8 @@ fun ProfileMainTab(
             ProfileMainContent(
                 profileImg = user.profileImageUrl,
                 nickname = user.nickname,
-                level = 4, /** 받을 수 있게 */
+                level = 4,
+                /** 받을 수 있게 */ /** 받을 수 있게 */
                 age = uiState.getAge(),
                 gender = user.gender,
                 introduction = user.description,
