+++ feature-mypage/src/main/java/com/materip/feature_mypage/screen/MyPage/ReveiwScreen.kt
@@ -1,4 +1,4 @@
-package com.materip.feature_mypage.screen
+package com.materip.feature_mypage.screen.MyPage
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -23,7 +23,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.materip.core_model.ui_model.ReviewClass
-import com.materip.feature_mypage.R
 import com.materip.matetrip.component.NormalTopBar
 import com.materip.matetrip.component.ReviewItem
