+++ feature-mypage/src/main/java/com/materip/feature_mypage/screen/MyPage/ReviewDescriptionScreen.kt
@@ -1,4 +1,4 @@
-package com.materip.feature_mypage.screen
+package com.materip.feature_mypage.screen.MyPage
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
@@ -13,7 +13,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
@@ -39,8 +38,6 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.materip.core_model.ui_model.ReviewDescClass
-import com.materip.feature_mypage.R
-import com.materip.matetrip.component.ImageLoadView
 import com.materip.matetrip.component.NormalTag
 import com.materip.matetrip.component.NormalTopBar
 import com.materip.matetrip.icon.Badges
